policy_hal: Fixed stream volume on Bluetooth Headset
When alarm tone is played during music playback on BT Headset,
music is stopped and resumed with wrong stream volume after
the alarm is stopped.
Sometimes startOutput for music comes before stopOutput of
alarm. AudioPolicy selects music volume according to combo
device in this case. When alarm is stopped later, music volume
is supposed to be restored according to headset but it is not
done. applyStreamVolumes is not called from setOutputDevice
because new device is same as previous device.
Fix is to call applyStreamVolumes explicitly from stopSource
whenever force is false

CRs-Fixed: 2058695

Change-Id: I12e0648512d7d2829285a7ca5ebec6152e90e4ee